FIX: Wrong value for Progressbar on large files + never ending downloads on large files

+++ src/de/snertlab/xdccBee/irc/DccDownload.java
@@ -107,8 +107,7 @@ public class DccDownload {
 
 		@Override
 		public void run() {
-			while ((int) dccFileTransfer.getProgress() < dccFileTransfer
-					.getSize()) {
+			while (dccFileTransfer.getProgress() < dccFileTransfer.getSize()) {
 				if (stop)
 					break;
 				tableItemDownload.getDisplay().asyncExec(new Runnable() {
